// src/shared/types/mock-server-data.type.ts
export type MockServerData = {
  categories: string[];
  titles: string[];
  descriptions: string[];
  cities: string[];
  previews: string[];
  images: string[];
  boolean: string[];
  housingTypes: string[];
  comforts: string[];
  authorNames: string[];
  authorEmails: string[];
  authorAvatars: string[];
};